Add setup.py and refactor game classes
The setup.py file was added to enable setup and installation of the project. The GameManager class in game.py was refactored into an abstract base class (Game) and derived class (TicTacToeGameManager). Node.py was updated to reflect changes in class hierarchy. Other logic for evaluating board states and generating board actions were also moved to the corresponding classes.